+++ Funcky.Test/Extensions/InterleaveTest.cs
@@ -70,5 +70,35 @@ namespace Funcky.Test.Extensions
 
             Assert.Equal(new List<int> { 1, 42, 1, 42 }, sequences.Interleave());
         }
+
+        [Fact]
+        public void GivenOneSequenceWithElementsAndAllTheOtherSequencesEmptyWeGetTheFirstSequence()
+        {
+            IEnumerable<int> sequence = new List<int> { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12 };
+            IEnumerable<int> emptySequence = Enumerable.Empty<int>();
+
+            Assert.Equal(sequence, emptySequence.Interleave(emptySequence, sequence, emptySequence));
+        }
+
+        [Fact]
+        public void GivenASequenceOfSequencesInterleaveReturnsTheExpectedSequence()
+        {
+            var sequences = new List<IEnumerable<int>>()
+            {
+                Enumerable.Repeat(1, 10),
+                Enumerable.Repeat(2, 10),
+                Enumerable.Repeat(3, 10),
+                Enumerable.Repeat(4, 10),
+            };
+
+            Assert.Equal(sequences.Select(s => s.Count()).Sum(), sequences.Interleave().Count());
+
+            int expected = 1;
+            foreach (var element in sequences.Interleave())
+            {
+                Assert.Equal(expected, element);
+                expected = (expected % 4) + 1;
+            }
+        }
     }
 }