+++ Funcky/Extensions/EnumerableExtensions/Interleave.cs
@@ -6,15 +6,29 @@ namespace Funcky.Extensions
 {
     public static partial class EnumerableExtensions
     {
-        public static IEnumerable<T> Interleave<T>(this IEnumerable<T> sequence, params IEnumerable<T>[] otherSequences)
+        /// <summary>
+        /// Interleaves the elements of multiple sequences by consuming the heads of each subsequence in the same order as the given subsequences. This repeats until all the sequences are completley consumed.
+        /// </summary>
+        /// <typeparam name="TSource">The type of the elements in the source sequences.</typeparam>
+        /// <param name="sequence">first sequence.</param>
+        /// <param name="otherSequences">other sequences.</param>
+        /// <returns>one sequences with all the elements interleaved.</returns>
+        [Pure]
+        public static IEnumerable<TSource> Interleave<TSource>(this IEnumerable<TSource> sequence, params IEnumerable<TSource>[] otherSequences)
             => Interleave(new[] { sequence }.Concat(otherSequences));
 
+        /// <summary>
+        /// Interleaves the elements of a sequence of sequences by consuming the heads of each subsequence in the same order as the given subsequences. This repeats until all the sequences are completley consumed.
+        /// </summary>
+        /// <typeparam name="TSource">The type of the elements in the source sequences.</typeparam>
+        /// <param name="source">the source sequences.</param>
+        /// <returns>one sequences with all the elements interleaved.</returns>
         [Pure]
         public static IEnumerable<TSource> Interleave<TSource>(this IEnumerable<IEnumerable<TSource>> source)
         {
-            List<IEnumerator<TSource>> sourceEnumerators = source.Select(s => s.GetEnumerator()).ToList();
+            var sourceEnumerators = source.Select(s => s.GetEnumerator()).ToList();
 
-            List<IEnumerator<TSource>> enumerators = sourceEnumerators;
+            var enumerators = sourceEnumerators;
             while (enumerators.Count > 0)
             {
                 enumerators = enumerators.Where(e => e.MoveNext()).ToList();