+++ changelog.md
@@ -100,3 +100,4 @@
 ## Unreleased
 * Funcky now uses `.ConfigureAwait(false)` everywhere `await` is used.
 * Add `WhereSelect`, `FirstOrNoneAsync`, `LastOrNoneAsync` and `SingleOrNoneAsync` extension methods for `IAsyncEnumerable`.
+* Add `Interleave`  extension methods for `IEnumerable`.